+++ src/helpers/wp-http.class.ts
@@ -1,0 +1,138 @@
+import {Injectable} from '@angular/core';
+import {Http, RequestOptions, RequestOptionsArgs, ConnectionBackend} from '@angular/http';
+import {Observable} from 'rxjs';
+
+import {ConfigService} from "../service/config/config.service";
+import {WpQueryArgs} from "./wp-query.class";
+
+@Injectable()
+export class WpHttp extends Http {
+
+  constructor(backend: ConnectionBackend,
+              defaultOptions: RequestOptions,
+              private _wpConfig: ConfigService) {
+    super(backend, defaultOptions);
+  }
+
+  direct(url: string): Observable<any> {
+
+    this.onStart();
+    return super.get(url, this.getOptions())
+
+      .timeout(this._wpConfig.timeOut, new Error('delay exceeded'))
+      .catch((err) => {
+        return this.onError(err);
+      })
+      .finally(() => {
+        this.onComplete();
+      });
+  }
+
+  get(endpoint: any, args?: WpQueryArgs): Observable<any> {
+
+    this.onStart();
+    return super.get(this.getUrl(endpoint, args), this.getOptions())
+
+      .timeout(this._wpConfig.timeOut, new Error('delay exceeded'))
+      .catch((err) => {
+        return this.onError(err);
+      })
+      .finally(() => {
+        this.onComplete()
+      });
+  }
+
+  post(endpoint: any, body: any): Observable<any> {
+
+    this.onStart();
+    return super.post(this.getUrl(endpoint), body, this.getOptions())
+
+      .timeout(this._wpConfig.timeOut, new Error('delay exceeded'))
+      .catch((err) => {
+        return this.onError(err);
+      })
+      .finally(() => {
+        this.onComplete();
+      });
+  }
+
+  put(endpoint: any, body: any): Observable<any> {
+
+    this.onStart();
+    return super.put(this.getUrl(endpoint), body, this.getOptions())
+
+      .timeout(this._wpConfig.timeOut, new Error('delay exceeded'))
+      .catch((err) => {
+        return this.onError(err);
+      })
+      .finally(() => {
+        this.onComplete();
+      });
+  }
+
+  delete(endpoint: any): Observable<any> {
+
+    this.onStart();
+    return super.delete(this.getUrl(endpoint), this.getOptions())
+
+      .timeout(this._wpConfig.timeOut, new Error('delay exceeded'))
+      .catch((err) => {
+        return this.onError(err);
+      })
+      .finally(() => {
+        this.onComplete();
+      });
+  }
+
+  /** Before the request  */
+  private onStart() {
+    /** set loading to true */
+    this._wpConfig.loading.next(true);
+  }
+
+  /** Handle errors */
+  private onError(err): Observable<any> {
+    /** Stream errors in wpConfig */
+    if (!(err.status === 400 || err.status === 422)) {
+
+    }
+    this._wpConfig.errors.next(err);
+    return Observable.empty();
+  }
+
+  /** Request completed */
+  private onComplete() {
+    /** set loading to false */
+    this._wpConfig.loading.next(false);
+  }
+
+  /** Get authKeys in headers for all requests */
+  private getOptions = (): RequestOptionsArgs => {
+    return {headers: this._wpConfig.getAuth()};
+  };
+
+  /** Serialize url endpoint and queryArgs */
+  private getUrl = (endpoint: string, args?: any): string => {
+    let url = this._wpConfig.baseUrl + endpoint;
+    if (args) {
+      /** add args to url */
+      url += '?' + serialize(args);
+    }
+    return url;
+  };
+
+}
+
+var serialize = (obj, prefix?): string => {
+  var str = [];
+  for (var p in obj) {
+    if (obj.hasOwnProperty(p) && obj[p]) {
+      var k = prefix ? prefix + "[" + p + "]" : p, v = obj[p];
+      str.push(typeof v == "object" ?
+        serialize(v, k) :
+      encodeURIComponent(k) + "=" + encodeURIComponent(v));
+    }
+  }
+  return str.join("&");
+};
+