+++ src/service/authentication/auth.interface.ts
@@ -1,5 +1,7 @@
-export interface IAuthenticate {
+import {Observable} from "rxjs/Observable";
+export interface AuthInterface {
 
-    basic(username: string, password: string): void;
-    logout(): void;
+  basic(username: string, password: string, remember?: boolean): Observable<any>;
+  cookies(): Observable<any>;
+  logout(): void;
 }