+++ src/service/authentication/auth.service.ts
@@ -1,49 +1,71 @@
 /**
  *  Authentication Service
- *  Basic Authenction
  */
 import {Injectable} from "@angular/core";
-import {Observable} from "rxjs";
+import {Observable} from "rxjs/Observable";
 
-import {IAuthenticate} from "./auth.interface";
-import {ConfigService} from "../config/config.service";
-import {WpHttp} from "../../classes/wp.http";
-import {WpHelper} from "../../classes/wp.helper";
-import {User} from "../../classes/user.model";
+import {AuthInterface} from "./auth.interface";
+import {ConfigService, AuthType} from "../config/config.service";
+import {WpEndpoint} from '../../helpers/wp-endpoints';
+import {WpHttp} from "../../helpers/wp-http.class";
 
 @Injectable()
-export class AuthService implements IAuthenticate {
+export class AuthService implements AuthInterface {
 
-    constructor(private _http: WpHttp, private _wpConfig: ConfigService) {
+  constructor(private _http: WpHttp, private _config: ConfigService) {
 
+  }
+
+  basic(username: string, password: string, remember?: boolean): Observable<any> {
+
+    /** Encode user keys */
+    let encodedCred = btoa(username + ':' + password);
+    return this.login(encodedCred, AuthType.basic).map(
+      (user)=> {
+        if (user) {
+          /** if login success, store key in localStorage */
+          if (remember) {
+            localStorage.setItem('authKeys', encodedCred);
+          }
+          else {
+            localStorage.removeItem('authKeys');
+          }
+          return user;
+        }
+      });
+  }
+
+  cookies(): Observable<any> {
+
+    let nonce = window['wpApiSettings'].nonce || undefined;
+    if (nonce) {
+      return this.login(nonce, AuthType.cookies);
     }
+  }
 
-    public logout = (): void=> {
-        this._wpConfig.authUser.next(null);
-    }
+  private login(keys: string, type: AuthType): Observable<any> {
 
-    public basic = (username: string, password: string, remember?: boolean): void => {
+    this._config.setAuth(keys, type);
 
-        /** Encode user keys */
-        let encodedCred = btoa(username + ':' + password);
+    return this._http.get(WpEndpoint.authentication).map(
+      (res)=> {
+        /** if login fail, send error message */
+        if (res && res.message) {
+          this._config.errors.next(res.message);
+          return undefined;
+        }
+        else {
+          this._config.setAuth(keys, type);
+          return res.json().body;
+        }
+      });
+  }
 
-        this.login(encodedCred).subscribe(
-            (user: User)=> {
-                /** send authUser */
-                this._wpConfig.authUser.next(user);
-                /** store keys for the current session */
-                this._wpConfig.authKeys = encodedCred;
-                /** store key in localStorage */
-                if (remember) {
-                    localStorage.setItem('authKeys', encodedCred);
-                } else {
-                    localStorage.removeItem('authKeys');
-                }
-            });
-    }
-
-    private login(encodedKeys: string): Observable<any> {
-        return this._http.post(WpHelper.endpoint.basicAuth, encodedKeys);
-    }
+  logout() {
+    this._config.setAuth(undefined, undefined);
+  }
 
 }
+
+
+