+++ src/service/collection/collection.service.ts
@@ -5,8 +5,8 @@
 import {Injectable} from "@angular/core";
 import {Observable} from "rxjs/Observable";
 
-import {QueryArgs} from "../../classes/args.model";
-import {WpHttp} from "../../classes/wp.http";
+import {WpQueryArgs} from "../../helpers/wp-query.class";
+import {WpHttp} from "../../helpers/wp-http.class";
 import {CollectionInterface} from "./collection.interface";
 import {Headers} from '@angular/http';
 
@@ -14,14 +14,14 @@ import {Headers} from '@angular/http';
 export class CollectionService implements CollectionInterface {
 
   /** Request Parameter  */
-  private queryArgs: QueryArgs;
+  private _args: WpQueryArgs;
 
   /** Collection Pagination Properties */
-  private _pagination: Pagination;
+  private _pagination: WpPagination;
   private _items = [];
 
   constructor(private http: WpHttp, private endpoint: string) {
-    this.queryArgs = new QueryArgs({});
+    this._args = new WpQueryArgs({});
   }
 
   /**
@@ -29,11 +29,11 @@ export class CollectionService implements CollectionInterface {
    * @param args
    * @returns {Observable<Response>}
    */
-  public get = (args?: QueryArgs): Observable<any> => {
+  public get = (args?: WpQueryArgs): Observable<any> => {
     /** reset pagination */
-    this._pagination = new Pagination();
+    this._pagination = new WpPagination();
     if (args) {
-      this.queryArgs = args;
+      this._args = args;
       /** if args.page is provided set the pagination currenPage */
       if (args.page) {
         this._pagination.currentPage = args.page;
@@ -49,15 +49,57 @@ export class CollectionService implements CollectionInterface {
   };
 
   /**
-   * Get next page of the collection
+   * Get next page of the collection combined with current collection
    * @returns {any}
    */
   public more = (): Observable<any> => {
-    if (this._pagination.hasMore) {
+    if (this._pagination && this._pagination.hasMore) {
       /** increment currentPage then set page argument */
-      this.queryArgs.page = ++this._pagination.currentPage;
+      this._args.page = ++this._pagination.currentPage;
       return this.fetch().map((res)=> {
         this._items = this._items.concat(res);
+        return {
+          data: this._items,
+          pagination: this._pagination
+        };
+      });
+    }
+    else {
+      return Observable.empty();
+    }
+  };
+
+  /**
+   * Get next page of the collection
+   * @returns {any}
+   */
+  public next = (): Observable<any> => {
+    if (this._pagination && this._pagination.hasMore) {
+      /** increment currentPage then set page argument */
+      this._args.page = ++this._pagination.currentPage;
+      return this.fetch().map((res)=> {
+        this._items = res;
+        return {
+          data: this._items,
+          pagination: this._pagination
+        };
+      });
+    }
+    else {
+      return Observable.empty();
+    }
+  };
+
+  /**
+   * Get prev page of the collection
+   * @returns {any}
+   */
+  public prev = (): Observable<any> => {
+    if (this._pagination && this._pagination.hasPrev) {
+      /** decrement currentPage then set page argument */
+      this._args.page = --this._pagination.currentPage;
+      return this.fetch().map((res)=> {
+        this._items = res;
         return {
           data: this._items,
           pagination: this._pagination
@@ -75,7 +117,7 @@ export class CollectionService implements CollectionInterface {
    */
   private fetch = (): Observable<any> => {
 
-    return this.http.get(this.endpoint, this.queryArgs).map(
+    return this.http.get(this.endpoint, this._args).map(
       (res) => {
         /** Set pagination  */
         this.setPagination(res.headers);
@@ -89,7 +131,7 @@ export class CollectionService implements CollectionInterface {
    * @param headers
    * @returns {Pagination}
    */
-  private setPagination = (headers: Headers): Pagination => {
+  private setPagination = (headers: Headers): WpPagination => {
 
     /** Fix issue of different property names in response headers */
     this._pagination.totalPages =
@@ -108,7 +150,7 @@ export class CollectionService implements CollectionInterface {
 /**
  * Pagination class holds the current collection response pagination and links
  */
-export class Pagination {
+export class WpPagination {
   /** Pagination holds the navigation data and links provided from WP API response header*/
   constructor(public currentPage: number = 1,
               public totalPages: number = 0,
@@ -119,5 +161,8 @@ export class Pagination {
 
   get hasMore(): boolean {
     return this.currentPage < this.totalPages;
+  }
+  get hasPrev(): boolean{
+    return this.currentPage > 1;
   }
 }