+++ src/service/config/config.interface.ts
@@ -1,11 +1,12 @@
-import {BehaviorSubject} from 'rxjs';
+import {Headers} from "@angular/http";
+import {BehaviorSubject} from 'rxjs/BehaviorSubject';
 
-export interface ConfigInterface{
+export interface ConfigInterface {
 
-    baseUrl:string;
-    authKeys:string;
-    timeOut: number;
-    loading:BehaviorSubject<boolean>;
-    errors: BehaviorSubject<any>;
-    authUser: BehaviorSubject<any>;
+  baseUrl: string;
+  timeOut: number;
+  loading: BehaviorSubject<boolean>;
+  errors: BehaviorSubject<any>;
+  getAuth(): Headers;
+  setAuth(keys,type);
 }