+++ src/service/model/model.interface.ts
@@ -1,9 +1,9 @@
 import {Observable} from 'rxjs/Observable';
-import {QueryArgs} from "../../classes/args.model";
+import {WpQueryArgs} from "../../helpers/wp-query.class";
 
 export interface ModelInterface {
 
-    get(id: number, args?: QueryArgs): Observable<any>;
+    get(id: number, args?: WpQueryArgs): Observable<any>;
     add(body: any): Observable<any>;
     update(id: number, body: any): Observable<any>;
     delete(id: number): Observable<any>;