+++ src/service/wp.interface.ts
@@ -3,10 +3,11 @@ import {EndpointService} from "./endpoint/endpoint.service";
 import {AuthService} from "./authentication/auth.service";
 
 export interface WpInterface {
-    link(url: string): Observable<any>;
-    collection(): EndpointService;
-    model(): EndpointService;
-    auth(): AuthService;
+  discover(url: string): Observable<any>;
+  link(url: string): Observable<any>;
+  collection(): EndpointService;
+  model(): EndpointService;
+  auth(): AuthService;
 }