+++ CHANGELOG.md
@@ -1,5 +1,47 @@
 #Changelog
 
+##2.1.0
+
+Use `WpService.config` instead of `WpConfig` to keep all usages in one service.
+    
+**Bugs**:
+    
+- Basic Authentication: Fixes request and return observable instead of void
+
+    `WpService.auth().basic(username,password).subscribe(...)`
+
+    
+**Features:**
+
+- CollectionService: Added `next` and `prev`
+
+  - `collectionService = wpService.collection().posts().get.subscribe(...)` Get collection
+  - `collectionService.more().subscrube(...)` Combine current page with next page collection
+  - `collectionService.next().subscrube(...)` Get next page collection
+  - `collectionService.prev().subscrube(...)` Get prev page collection
+
+    
+- Discovery endpoint: Discover if a url has WP API install.
+`WpService.discover(url).subscribe(...)`
+
+- Cookies Authentication
+`WpService.auth().cookies().subscribe(...)`
+
+Breaking Changes:
+
+All helper classes has been prefixed with `Wp`
+- `Post` is now `WpPost`
+- `User` is now `WpUser`
+- `QueryArgs` is now `WpQueryArgs`
+- `WpHelper.endpoint` is now `WpEndpoint`
+- `WpHelper` is deprecated.
+- `authentication` notifier is deprecated from `WpConfig`.
+
+##2.0.1
+
+No changes.
+- added keywords to `package.json`
+
 ##2.0.0
 
 The service has been rewritten from scratch